Box unfold mission
fixed contours, current approach:
blur image extensively > get the color of the center of the box and take as white > mask this color > use canny edge detection on the mask > blur the edges > get contours > approximate contour to rectangle > get rectangle edges > apply perspective correction and crop to the edges.

fixed finding the top side:
compared the bottom color to the center color instead of perfect white.

resized the images to match each other and fit the screen

placed the top box after the first one to match the mission requirements

>> colors still not matching accurately
>> normalizing the final image brightness could probably make it look better